requirements.txt
google-genai
pdfplumber
python-docx
streamlit
python-dotenv
